Moves movieInfo function (where it munges data) from rental controller into the rental model.
